requirements.txt
streamlit
langchain
langchain-openai
python-dotenv
pandas
pydantic
chromadb